requirements-dev.txt
# Development dependencies
# Include production dependencies
-r requirements.txt

# Testing
coverage

# Linting and formatting
flake8
black
isort
mypy

# Django extensions (optional)
django-extensions
